+++ lib/src/const.dart
@@ -1,0 +1,5 @@
+part of wire;
+
+final String ERROR__MIDDLEWARE_EXISTS =
+    'Middleware already registered, middleware: ';
+final String ERROR__WIRE_ALREADY_REGISTERED = 'Wire already registered, hash: ';